requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
argilla==1.14.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.7
clickhouse-connect==0.6.8
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.6
cryptography==41.0.3
dataclasses-json==0.5.14
Deprecated==1.2.14
diskcache==5.6.1
duckdb==0.8.1
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
extract-msg==0.41.5
fastapi==0.101.1
filelock==3.12.2
filetype==1.2.0
Flask==2.3.3
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
gpt4all==1.0.3
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
IMAPClient==2.3.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.228
langchainplus-sdk==0.0.20
lark-parser==0.12.0
llama-cpp-python==0.1.68
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.1.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.23.5
olefile==0.46
oletools==0.60.1
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandoc==2.3
pcodedmp==1.2.6
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
plumbum==1.8.2
ply==3.11
posthog==3.0.2
protobuf==4.24.1
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyMuPDF==1.22.5
pypandoc==1.11
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
red-black-tree-mod==1.20
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.5.2
RTFDE==0.0.2
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
unstructured==0.8.0
urllib3==2.0.3
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
